Use single redis protocol
Move it to it's own file
